[OSDEV-2786] Move anonymise_in_paid_products from User to Contributor
The flag controls whether a *contributor's* identity is masked in paid
products, so it belongs on Contributor, not on the admin User account.

- Add `anonymise_in_paid_products` (BooleanField, default False) to
  `Contributor` and its `HistoricalContributor` mirror (migration 0218).
- Override `Contributor.save()` to detect flag changes and flush
  `view_cache` immediately, matching the previous behaviour on User.
- Remove the field and the cache-invalidation logic from `User`.
- Update `admin.py`: remove from `OarUserAdmin`, expose in
  `ContributorAdmin` with an explicit `fields` list.
- Update `ShouldMaskContribution._load()` to filter on
  `anonymise_in_paid_products` directly instead of via
  `admin__anonymise_in_paid_products`.
- Fix a pre-existing crash in `_format_source`: `last_contributed_at`
  now guards `None` the same way `list_uploaded_at` already did, so
  the anonymous "Other" merged entry (which carries no date) no longer
  raises TypeError.
- Update all tests to set the flag on Contributor instead of User.
